Move !join and friends to a new module
